Better error and issue handling -Added return value for most functions in VkFFT with error handling on all memory allocations -Fixed most compile warnings (#38 and some others) -Better resources management and deallocation on failure

+++ benchmark_scripts/vkFFT_scripts/include/utils_VkFFT.h
@@ -61,5 +61,5 @@ VkFFTResult transferDataToCPU(VkGPU* vkGPU, void* arr, VkBuffer* buffer, uint64_
 #endif
 VkFFTResult devices_list();
 VkFFTResult performVulkanFFT(VkGPU* vkGPU, VkFFTApplication* app, VkFFTLaunchParams* launchParams, int inverse, uint64_t num_iter);
-VkFFTResult performVulkanFFTiFFT(VkGPU* vkGPU, VkFFTApplication* app, VkFFTLaunchParams* launchParams, uint64_t num_iter, float* time_result);
+VkFFTResult performVulkanFFTiFFT(VkGPU* vkGPU, VkFFTApplication* app, VkFFTLaunchParams* launchParams, uint64_t num_iter, double* time_result);
 #endif
